detect modified config files by checking content rather than mtime
Closes #19.

In the very beginning, Holo detected user modifications on configuration
files by checking the file content. This was removed in v0.2 (or, to be
more exact, in 7a4d65c) when holoscripts were introduced.

However, the mtime check that was introduced has recently proven
unstable in unit tests and hard to test correctly (which is why it is
not tested at all right now). The new approach is to save a copy of each
generated configuration file during the application algorithm, in the
new /var/lib/holo/computed directory, and compare the contents of this
file (or symlink) to the target prior to attempting to overwrite it.