v0.6.0 "Providence"
Starting with this release, I'll choose a codename for every feature
release, sometimes with the intention of depicting a theme for this
release's development cycle.

This release is called "Providence" because the new `holo diff`
subcommand allows system administrators to act more prudently in the
resolution of manual changes to configuration targets.

The next release will be called "Harmony" because it will focus on
making the interface as well as the mental model of Holo more
consistent, making all the features feel more like a unified whole.