optimize application algorithm through lazy evaluation
When a repository entry is a plain file, the results of previous
application steps are discarded. Therefore, we start evaluation at the
latest plain repositry file, skipping all previous application steps
entirely.

The test uses a failing holoscript followed by a plain repository file,
which causes the holoscript to never be executed.

Closes #44.